+++ packages/web-app/src/utils/contract.ts
@@ -1,5 +1,8 @@
 // This file has been generated by james but committed by me
 
+import {AugmentedEtherscanContractResponse} from 'containers/smartContractComposer/components/contractAddressValidation';
+import {SmartContractAction} from './types';
+
 export interface NatspecDetails {
   keyword: string;
   name: string;
@@ -289,3 +292,21 @@ export const flattenNatSpecTags = (
 
   return flatTags;
 };
+
+export function attachEtherNotice(
+  SourceCode: AugmentedEtherscanContractResponse['SourceCode'],
+  ContractName: string,
+  ABI: SmartContractAction[]
+): SmartContractAction[] {
+  const parsedSourceCode = parseSourceCode(SourceCode);
+  const EtherNotice = extractNatSpec(parsedSourceCode);
+  const notices = EtherNotice[ContractName]?.details;
+
+  return ABI.map(action => {
+    if (action.type === 'function' && notices?.[action.name]) {
+      action.notice = notices[action.name].tags.notice as string;
+    }
+
+    return action;
+  });
+}
